+++ src/django/VLE/old_views/get.py
@@ -23,19 +23,6 @@ import VLE.serializers as serialize
 import VLE.permissions as permissions
 import VLE.views.responses as responses
 
-# VUE ENTRY STATE
-BAD_AUTH = '-1'
-
-NO_USER = '0'
-LOGGED_IN = '1'
-
-NO_COURSE = '0'
-NO_ASSIGN = '1'
-NEW_COURSE = '2'
-NEW_ASSIGN = '3'
-FINISH_T = '4'
-FINISH_S = '5'
-GRADE_CENTER = '6'
 
 
 # @api_view(['GET'])
@@ -171,114 +158,3 @@ GRADE_CENTER = '6'
 #         return responses.forbidden('You are not allowed to edit the courses.')
 #
 #     return responses.success(payload={'assignment': serialize.assignment_to_dict(assignment)})
-
-
-@api_view(['GET'])
-def get_lti_params_from_jwt(request, jwt_params):
-    """Handle the controlflow for course/assignment create, connect and select.
-
-    Returns the data needed for the correct entry place.
-    """
-    if not request.user.is_authenticated:
-        return responses.unauthorized()
-
-    user = request.user
-    lti_params = jwt.decode(jwt_params, settings.LTI_SECRET, algorithms=['HS256'])
-    roles = json.load(open('config.json'))
-    lti_roles = dict((roles[k], k) for k in roles)
-    role = lti_roles[lti_params['roles']]
-
-    payload = dict()
-    course = lti.check_course_lti(lti_params, user, role)
-    if course is None:
-        if role == 'Teacher':
-            payload['state'] = NEW_COURSE
-            payload['lti_cName'] = lti_params['context_title']
-            payload['lti_abbr'] = lti_params['context_label']
-            payload['lti_cID'] = lti_params['context_id']
-            payload['lti_aName'] = lti_params['resource_link_title']
-            payload['lti_aID'] = lti_params['resource_link_id']
-
-            if 'custom_canvas_assignment_points_possible' in lti_params:
-                payload['lti_points_possible'] = lti_params['custom_canvas_assignment_points_possible']
-
-            return responses.success(payload={'params': payload})
-        else:
-            return responses.not_found(description='The assignment you are looking for cannot be found. \
-                <br>Note: it might still be reachable through the assignment section')
-
-    assignment = lti.check_assignment_lti(lti_params)
-    if assignment is None:
-        if role == 'Teacher':
-            payload['state'] = NEW_ASSIGN
-            payload['cID'] = course.pk
-            payload['lti_aName'] = lti_params['resource_link_title']
-            payload['lti_aID'] = lti_params['resource_link_id']
-
-            if 'custom_canvas_assignment_points_possible' in lti_params:
-                payload['lti_points_possible'] = lti_params['custom_canvas_assignment_points_possible']
-
-            return responses.success(payload={'params': payload})
-        else:
-            return responses.not_found(description='The assignment you are looking for cannot be found. \
-                <br>Note: it might still be reachable through the assignment section')
-
-    journal = lti.select_create_journal(lti_params, user, assignment, roles)
-    jID = journal.pk if journal is not None else None
-    state = FINISH_T if jID is None else FINISH_S
-
-    payload['state'] = state
-    payload['cID'] = course.pk
-    payload['aID'] = assignment.pk
-    payload['jID'] = jID
-    return responses.success(payload={'params': payload})
-
-
-@api_view(['POST'])
-def lti_launch(request):
-    """Django view for the lti post request.
-
-    handles the users login or sned to a creation page.
-    """
-    secret = settings.LTI_SECRET
-    key = settings.LTI_KEY
-
-    authenticated, err = lti.OAuthRequestValidater.check_signature(
-        key, secret, request)
-
-    if authenticated:
-        roles = json.load(open('config.json'))
-        params = request.POST.dict()
-        user = lti.check_user_lti(params, roles)
-
-        params['exp'] = datetime.datetime.utcnow() + datetime.timedelta(minutes=15)
-        lti_params = jwt.encode(params, secret, algorithm='HS256').decode('utf-8')
-
-        if user is None:
-            q_names = ['state', 'lti_params']
-            q_values = [NO_USER, lti_params]
-
-            if 'lis_person_name_full' in params:
-                fullname = params['lis_person_name_full']
-                splitname = fullname.split(' ')
-                firstname = splitname[0]
-                lastname = fullname[len(splitname[0])+1:]
-                q_names += ['firstname', 'lastname']
-                q_values += [firstname, lastname]
-
-            if 'lis_person_sourcedid' in params:
-                q_names.append('username')
-                q_values.append(params['lis_person_sourcedid'])
-
-            if 'lis_person_contact_email_primary' in params:
-                q_names.append('email')
-                q_values.append(params['lis_person_contact_email_primary'])
-
-            return redirect(lti.create_lti_query_link(q_names, q_values))
-
-        refresh = TokenObtainPairSerializer.get_token(user)
-        access = refresh.access_token
-        return redirect(lti.create_lti_query_link(['lti_params', 'jwt_access', 'jwt_refresh', 'state'],
-                                                  [lti_params, access, refresh, LOGGED_IN]))
-
-    return redirect(lti.create_lti_query_link(['state'], [BAD_AUTH]))
